Fix file path regex pattern in stack trace response
The original regex pattern would encounter for characters other than white space (\S)
as part of the file path. In some cases this is not true.
Apparently a space is a valid file path character and must be encountered for.

This bug would prevent the VSCode from stopping on break points, whereas
the actual Perl debugger would indicate a break point was hit.